requirements.txt
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Werkzeug==2.2.2
SQLAlchemy==2.0.3
